+++ src/App.tsx
@@ -3,6 +3,7 @@ import Layout from './components/Layout'
 import Setting from './pages/Setting'
 import Ready from './pages/Ready'
 import Queue from './pages/Queue'
+import Zone from './pages/Zone'
 import Seat from './pages/Seat'
 import Captcha from './pages/Captcha'
 import Result from './pages/Result'
@@ -15,6 +16,7 @@ const router = createHashRouter([
       { path: '/ready', element: <Ready /> },
       { path: '/queue', element: <Queue /> },
       { path: '/captcha', element: <Captcha /> },
+      { path: '/zone', element: <Zone /> },
       { path: '/seat', element: <Seat /> },
       { path: '/result', element: <Result /> },
     ],